requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==1.2.2
pydantic==2.12.5
langchain==1.2.11
langchain-core==1.2.18
langchain-google-genai==4.2.1
langchain-chroma==1.1.0
langchain-text-splitters==1.1.1
langgraph==1.1.0
chromadb==1.5.5
google-generativeai==0.8.6
google-genai==1.66.0
